+++ src/lib/components/BodySidePanel.svelte
@@ -93,7 +93,7 @@
     {:else if selectedTab === 'Orbit'}
       <BodyOrbitTab {body} {parentBody} {system} {rulePack} on:update={handleUpdate} />
     {:else if selectedTab === 'Temp'}
-      <BodyTemperatureTab {body} {rulePack} {rootStar} {parentBody} nodes={system.nodes} on:update={handleUpdate} />
+      <BodyTemperatureTab {body} {rulePack} {rootStar} {parentBody} nodes={system.nodes} systemAgeGyr={(system as any).age_Gyr ?? 4.6} on:update={handleUpdate} />
     {:else if selectedTab === 'Atmosphere'}
       <BodyAtmosphereTab {body} {rulePack} {system} on:update={handleUpdate} />
     {:else if selectedTab === 'Hydro'}